python/requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
typing-extensions>=4.0.0
pydantic>=2.0.0
httpx>=0.24.0
aiohttp>=3.8.0
urllib3>=2.0.0
certifi>=2023.0.0
chardet>=5.0.0
idna>=3.4
PySocks>=1.7.1 